Audit follow-up: fix verified bugs, add prefix-hash markers, repair rendering, modernize stress tests
- preprocessor: '// <hash>' markers print the hash of the listing up to
  that line; header.tmp cleanup paths fixed; long page headers shrink via
  resizebox (old fontsize had no selectfont and never worked); listing
  captions share the description line only when they fit (no more overlap)
- fix confirmed content bugs (Trie merge/mex/lazy pushes, Treap del,
  LIS dp bounds, Blossom edge-0/cnt/init, SolveLinearBinary, Simplex,
  BigFloat division, RNG Mersenne mask, BeamSearch guards, FactorSqrt
  overflow, HillClimbing typo, KMP pi rename, WPM rewrite on convMod,
  GomoryHu/DirectedMST/LCA rewired to current APIs, hungarian ll,
  GlobalMinCut ll, FloydWarshall/BellmanFord on template INF, etc)
- fix prose errors (Dilworth, min-cut reduction orientation, Hall's,
  Faulhaber, Stirling row, Normal mean, Markov conventions, divisor
  table 2304, Green Hackenbush, markdown bold)
- interactor.py runnable + correct exit codes; stress.sh bash shebang
- stress-tests: harness template covers both macro dialects; stale tests
  rewritten against current APIs; new tests for previously untested
  structures (all pass); doc/audit-notes.md records conventions/landmines

+++ content/geometry/Angle.h
@@ -15,9 +15,9 @@
 #pragma once
 
 struct Angle {
-	int x, y;
+	ll x, y;
 	int t;
-	Angle(int x, int y, int t = 0) : x(x), y(y), t(t) {}
+	Angle(ll x, ll y, int t = 0) : x(x), y(y), t(t) {}
 	Angle operator-(Angle b) const { return {x - b.x, y - b.y, t}; }
 	int half() const {
 		assert(x || y);
@@ -29,8 +29,8 @@ struct Angle {
 };
 bool operator<(Angle a, Angle b) {
 	// add a.dist2() and b.dist2() to also compare distances
-	return make_tuple(a.t, a.half(), a.y * (ll) b.x) <
-	       make_tuple(b.t, b.half(), a.x * (ll) b.y);
+	return make_tuple(a.t, a.half(), a.y * b.x) <
+	       make_tuple(b.t, b.half(), a.x * b.y);
 }
 
 // Given two points, this calculates the smallest angle between
